Delete test_perfilter_internal.py
Looking over this file, I don't get the sense that it's needed.
perfilter_internal is only ever run when mif is run, and mif is already tested
in the files that test fit. Many of the mif tests appear to be the same as
those in the perfilter tests. Furthermore, deleting test_perfilter_internal.py
does not affect the test coverage. Consequently, I think this should be
deleted. If a problem arises, we can resurrect the file.